+++ PolyvVodSDKDemo/PolyvVodSDKDemo/PolyvOpenSourceModule/Skin/View/PLVVodShrinkscreenView.h
@@ -22,6 +22,16 @@
 @property (weak, nonatomic) IBOutlet UIButton *videoPlayModeButton;
 @property (weak, nonatomic) IBOutlet UIButton *audioPlayModeButton;
 
+@property (weak, nonatomic) IBOutlet UIButton *definitionButton;
+@property (weak, nonatomic) IBOutlet UIButton *playbackRateButton;
+@property (weak, nonatomic) IBOutlet UIButton *routeButton; // 线路切换
+
+@property (nonatomic, assign) BOOL isShowRouteline; // 显示线路
+@property (nonatomic, assign) BOOL isShowRate;      // 显示速率
+@property (nonatomic, assign) BOOL isShowQuality;   // 显示清晰度
+
+@property (nonatomic, assign) BOOL enableQualityBtn;   // 清晰度按钮是否响应事件
+
 - (void)switchToPlayMode:(PLVVodPlaybackMode)mode;
 
 @end